xds: move e2e tests into grpc/test/xds directory
